Do not let slice spacing make the chart go haywire. (Fixes #1582)
If the user provides data that does not allow the spacing to stay even -
he will need to manually fix that (i.e by giving a larger inner hole)

Even with a hole - with very small slices it will not help.
Spacing could be reduces, and small slices can be removed from the dataset.